Fixed issue with not being able to delete beer images but instead deleting all tap images

+++ admin/image_remove.php
@@ -1,5 +1,7 @@
 <?php
-$imgs = glob (  __DIR__.'/../img/tap/tap'.$_GET['tapId'].'*' );
+if( !isset($_GET['type']) || !isset($_GET['id']))exit;
+if( $_GET['type'] != "tap" && $_GET['type'] != "beer")exit;
+$imgs = glob (  __DIR__.'/../img/'.$_GET['type'].'/'.$_GET['type'].$_GET['id'].'*' );
 foreach( $imgs as $img ){
     unlink($img);
 }
